TIP-843: Move job parameters to enrichment
