+++ src/Akeneo/Pim/Enrichment/Bundle/DependencyInjection/AkeneoPimEnrichmentExtension.php
@@ -81,6 +81,7 @@ class AkeneoPimEnrichmentExtension extends Extension
         $loader->load('normalizers.yml');
         $loader->load('connector.yml');
         $loader->load('array_converter.yml');
+        $loader->load('job_parameters.yml');
 
         if (!$container->hasParameter('pim_pdf_generator_font')) {
             $container->setParameter('pim_pdf_generator_font', null);
