+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/job_parameters.yml
@@ -1,0 +1,72 @@
+services:
+    pim_connector.job.job_parameters.constraint_collection_provider.product_csv_export:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductCsvExport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_csv_export'
+            -
+                - 'csv_product_export'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_xlsx_export:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductXlsxExport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_xlsx_export'
+            -
+                - 'xlsx_product_export'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_model_csv_export:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductModelCsvExport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_csv_export'
+            -
+                - 'csv_product_model_export'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_model_xlsx_export:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductModelCsvExport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_xlsx_export'
+            -
+                - 'xlsx_product_model_export'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_csv_import:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductCsvImport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_csv_import'
+            -
+                - 'csv_product_import'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_model_csv_import:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductModelCsvImport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_csv_import'
+            -
+                - 'csv_product_model_import'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_xlsx_import:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductCsvImport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_xlsx_import'
+            -
+                - 'xlsx_product_import'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
+
+    pim_connector.job.job_parameters.constraint_collection_provider.product_model_xlsx_import:
+        class: 'Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductModelCsvImport'
+        arguments:
+            - '@pim_connector.job.job_parameters.constraint_collection_provider.simple_xlsx_import'
+            -
+                - 'xlsx_product_model_import'
+        tags:
+            - { name: akeneo_batch.job.job_parameters.constraint_collection_provider }
