+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Job/JobParameters/ConstraintCollectionProvider/ProductModelCsvImport.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Component\Connector\Job\JobParameters\ConstraintCollectionProvider;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
@@ -9,19 +9,18 @@ use Symfony\Component\Validator\Constraints\NotBlank;
 use Symfony\Component\Validator\Constraints\Type;
 
 /**
- * Constraints for product CSV import
  *
- * @author    Nicolas Dupont <nicolas@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @author    Arnaud Langlade <arnaud.langlade@akeneo.com>
+ * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductCsvImport implements ConstraintCollectionProviderInterface
+class ProductModelCsvImport implements ConstraintCollectionProviderInterface
 {
     /** @var ConstraintCollectionProviderInterface */
-    protected $simpleProvider;
+    private $simpleProvider;
 
     /** @var array */
-    protected $supportedJobNames;
+    private $supportedJobNames;
 
     /**
      * @param ConstraintCollectionProviderInterface $simpleProvider
@@ -36,7 +35,7 @@ class ProductCsvImport implements ConstraintCollectionProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getConstraintCollection()
+    public function getConstraintCollection(): Collection
     {
         $baseConstraint = $this->simpleProvider->getConstraintCollection();
         $constraintFields = $baseConstraint->fields;
@@ -44,8 +43,7 @@ class ProductCsvImport implements ConstraintCollectionProviderInterface
         $constraintFields['dateFormat'] = new NotBlank();
         $constraintFields['enabled'] = new Type('bool');
         $constraintFields['categoriesColumn'] = new NotBlank();
-        $constraintFields['familyColumn'] = new NotBlank();
-        $constraintFields['groupsColumn'] = new NotBlank();
+        $constraintFields['familyVariantColumn'] = new NotBlank();
         $constraintFields['enabledComparison'] = new Type('bool');
         $constraintFields['realTimeVersioning'] = new Type('bool');
         $constraintFields['invalid_items_file_format'] = new NotBlank();
@@ -56,7 +54,7 @@ class ProductCsvImport implements ConstraintCollectionProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function supports(JobInterface $job)
+    public function supports(JobInterface $job): bool
     {
         return in_array($job->getName(), $this->supportedJobNames);
     }
