+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Job/JobParameters/DefaultValueProvider/ProductCsvImport.php
@@ -1,24 +1,25 @@
 <?php
 
-namespace Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider;
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\DefaultValueProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\DefaultValuesProviderInterface;
 use Akeneo\Tool\Component\Localization\Localizer\LocalizerInterface;
 
 /**
+ * DefaultParameters for product CSV import
  *
- * @author    Arnaud Langlade <arnaud.langlade@akeneo.com>
- * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
+ * @author    Nicolas Dupont <nicolas@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductModelCsvImport implements DefaultValuesProviderInterface
+class ProductCsvImport implements DefaultValuesProviderInterface
 {
     /** @var DefaultValuesProviderInterface */
-    private $simpleProvider;
+    protected $simpleProvider;
 
     /** @var array */
-    private $supportedJobNames;
+    protected $supportedJobNames;
 
     /**
      * @param DefaultValuesProviderInterface $simpleProvider
@@ -33,14 +34,15 @@ class ProductModelCsvImport implements DefaultValuesProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getDefaultValues(): array
+    public function getDefaultValues()
     {
         $parameters = $this->simpleProvider->getDefaultValues();
         $parameters['decimalSeparator'] = LocalizerInterface::DEFAULT_DECIMAL_SEPARATOR;
         $parameters['dateFormat'] = LocalizerInterface::DEFAULT_DATE_FORMAT;
         $parameters['enabled'] = true;
         $parameters['categoriesColumn'] = 'categories';
-        $parameters['familyVariantColumn'] = 'family_variant';
+        $parameters['familyColumn'] = 'family';
+        $parameters['groupsColumn'] = 'groups';
         $parameters['enabledComparison'] = true;
         $parameters['realTimeVersioning'] = true;
 
@@ -50,7 +52,7 @@ class ProductModelCsvImport implements DefaultValuesProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function supports(JobInterface $job): bool
+    public function supports(JobInterface $job)
     {
         return in_array($job->getName(), $this->supportedJobNames);
     }
