+++ src/Akeneo/Pim/Enrichment/Component/Product/Connector/Job/JobParameters/DefaultValueProvider/ProductModelCsvExport.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider;
+declare(strict_types=1);
+
+namespace Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\DefaultValueProvider;
 
 use Akeneo\Channel\Component\Repository\ChannelRepositoryInterface;
 use Akeneo\Channel\Component\Repository\LocaleRepositoryInterface;
@@ -10,25 +12,25 @@ use Akeneo\Tool\Component\Localization\Localizer\LocalizerInterface;
 use Pim\Component\Catalog\Query\Filter\Operators;
 
 /**
- * DefaultParameters for product XLSX export
+ * DefaultParameters for product model CSV export
  *
- * @author    Marie Bochu <marie.bochu@akeneo.com>
- * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @author    Nicolas Dupont <nicolas@akeneo.com>
+ * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductXlsxExport implements DefaultValuesProviderInterface
+class ProductModelCsvExport implements DefaultValuesProviderInterface
 {
     /** @var DefaultValuesProviderInterface */
-    protected $simpleProvider;
-
-    /** @var ChannelRepositoryInterface */
-    protected $channelRepository;
-
-    /** @var LocaleRepositoryInterface */
-    protected $localeRepository;
+    private $simpleProvider;
 
     /** @var array */
-    protected $supportedJobNames;
+    private $supportedJobNames;
+
+    /** @var ChannelRepositoryInterface */
+    private $channelRepository;
+
+    /** @var LocaleRepositoryInterface */
+    private $localeRepository;
 
     /**
      * @param DefaultValuesProviderInterface $simpleProvider
@@ -42,9 +44,9 @@ class ProductXlsxExport implements DefaultValuesProviderInterface
         LocaleRepositoryInterface $localeRepository,
         array $supportedJobNames
     ) {
-        $this->simpleProvider = $simpleProvider;
+        $this->simpleProvider    = $simpleProvider;
         $this->channelRepository = $channelRepository;
-        $this->localeRepository = $localeRepository;
+        $this->localeRepository  = $localeRepository;
         $this->supportedJobNames = $supportedJobNames;
     }
 
@@ -57,7 +59,6 @@ class ProductXlsxExport implements DefaultValuesProviderInterface
         $parameters['decimalSeparator'] = LocalizerInterface::DEFAULT_DECIMAL_SEPARATOR;
         $parameters['dateFormat'] = LocalizerInterface::DEFAULT_DATE_FORMAT;
         $parameters['with_media'] = true;
-        $parameters['linesPerFile'] = 10000;
 
         $channels = $this->channelRepository->getFullChannels();
         $defaultChannelCode = (0 !== count($channels)) ? $channels[0]->getCode() : null;
@@ -68,13 +69,8 @@ class ProductXlsxExport implements DefaultValuesProviderInterface
         $parameters['filters'] = [
             'data'      => [
                 [
-                    'field'    => 'enabled',
-                    'operator' => Operators::EQUALS,
-                    'value'    => true,
-                ],
-                [
                     'field'    => 'completeness',
-                    'operator' => Operators::GREATER_OR_EQUAL_THAN,
+                    'operator' => Operators::AT_LEAST_COMPLETE,
                     'value'    => 100,
                 ],
                 [
