+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Job/JobParameters/ConstraintCollectionProvider/ProductCsvImportSpec.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Job\JobParameters\ConstraintCollectionProvider;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\Validator\Constraints\Collection;
 
-class ProductXlsxExportSpec extends ObjectBehavior
+class ProductCsvImportSpec extends ObjectBehavior
 {
     function let(ConstraintCollectionProviderInterface $decoratedProvider)
     {
@@ -27,9 +27,16 @@ class ProductXlsxExportSpec extends ObjectBehavior
         $collection = $this->getConstraintCollection();
         $collection->shouldReturnAnInstanceOf('Symfony\Component\Validator\Constraints\Collection');
         $fields = $collection->fields;
+        $fields->shouldHaveCount(9);
         $fields->shouldHaveKey('decimalSeparator');
         $fields->shouldHaveKey('dateFormat');
-        $fields->shouldHaveKey('filters');
+        $fields->shouldHaveKey('enabled');
+        $fields->shouldHaveKey('categoriesColumn');
+        $fields->shouldHaveKey('familyColumn');
+        $fields->shouldHaveKey('groupsColumn');
+        $fields->shouldHaveKey('enabledComparison');
+        $fields->shouldHaveKey('realTimeVersioning');
+        $fields->shouldHaveKey('invalid_items_file_format');
     }
 
     function it_supports_a_job(JobInterface $job)
