+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Job/JobParameters/ConstraintCollectionProvider/ProductModelCsvExportSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Job\JobParameters\ConstraintCollectionProvider;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
