+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Job/JobParameters/ConstraintCollectionProvider/ProductModelCsvImportSpec.php
@@ -1,22 +1,30 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Job\JobParameters\ConstraintCollectionProvider;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
-use PhpSpec\ObjectBehavior;
 use Symfony\Component\Validator\Constraints\Collection;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider\ProductModelCsvImport;
+use PhpSpec\ObjectBehavior;
+use Prophecy\Argument;
 
-class ProductCsvImportSpec extends ObjectBehavior
+class ProductModelCsvImportSpec extends ObjectBehavior
 {
     function let(ConstraintCollectionProviderInterface $decoratedProvider)
     {
         $this->beConstructedWith($decoratedProvider, ['my_supported_job_name']);
     }
 
+    function it_is_initializable()
+    {
+        $this->shouldHaveType(
+           ProductModelCsvImport::class);
+    }
+
     function it_is_a_provider()
     {
-        $this->shouldImplement('Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface');
+        $this->shouldImplement(ConstraintCollectionProviderInterface::class);
     }
 
     function it_provides_constraints_collection(
@@ -25,18 +33,17 @@ class ProductCsvImportSpec extends ObjectBehavior
     ) {
         $decoratedProvider->getConstraintCollection()->willReturn($decoratedCollection);
         $collection = $this->getConstraintCollection();
-        $collection->shouldReturnAnInstanceOf('Symfony\Component\Validator\Constraints\Collection');
+        $collection->shouldReturnAnInstanceOf(Collection::class);
         $fields = $collection->fields;
-        $fields->shouldHaveCount(9);
+        $fields->shouldHaveCount(8);
         $fields->shouldHaveKey('decimalSeparator');
         $fields->shouldHaveKey('dateFormat');
         $fields->shouldHaveKey('enabled');
         $fields->shouldHaveKey('categoriesColumn');
-        $fields->shouldHaveKey('familyColumn');
-        $fields->shouldHaveKey('groupsColumn');
         $fields->shouldHaveKey('enabledComparison');
         $fields->shouldHaveKey('realTimeVersioning');
         $fields->shouldHaveKey('invalid_items_file_format');
+        $fields->shouldHaveKey('familyVariantColumn');
     }
 
     function it_supports_a_job(JobInterface $job)
