+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Job/JobParameters/ConstraintCollectionProvider/ProductXlsxExportSpec.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Job\JobParameters\ConstraintCollectionProvider;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\ConstraintCollectionProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\ConstraintCollectionProviderInterface;
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\Validator\Constraints\Collection;
 
-class ProductCsvExportSpec extends ObjectBehavior
+class ProductXlsxExportSpec extends ObjectBehavior
 {
     function let(ConstraintCollectionProviderInterface $decoratedProvider)
     {
