+++ src/Akeneo/Pim/Enrichment/Component/spec/Product/Connector/Job/JobParameters/DefaultValueProvider/ProductModelCsvImportSpec.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace spec\Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider;
+namespace spec\Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\DefaultValueProvider;
 
 use Akeneo\Tool\Component\Batch\Job\JobInterface;
 use Akeneo\Tool\Component\Batch\Job\JobParameters\DefaultValuesProviderInterface;
-use Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModelCsvImport;
+use Akeneo\Pim\Enrichment\Component\Product\Connector\Job\JobParameters\DefaultValueProvider\ProductModelCsvImport;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
@@ -17,7 +17,8 @@ class ProductModelCsvImportSpec extends ObjectBehavior
 
     function it_is_initializable()
     {
-        $this->shouldHaveType(ProductModelCsvImport::class);
+        $this->shouldHaveType(
+            ProductModelCsvImport::class);
     }
 
     function it_is_a_provider()
