+++ src/Pim/Bundle/ConnectorBundle/Resources/config/job_defaults.yml
@@ -2,18 +2,10 @@ parameters:
     pim_connector.job.job_parameters.default_values_provider.simple_csv_export.class:         Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleCsvExport
     pim_connector.job.job_parameters.default_values_provider.simple_xlsx_export.class:        Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleXlsxExport
     pim_connector.job.job_parameters.default_values_provider.simple_yaml_export.class:        Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleYamlExport
-    pim_connector.job.job_parameters.default_values_provider.product_csv_export.class:        Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductCsvExport
-    pim_connector.job.job_parameters.default_values_provider.product_xlsx_export.class:       Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductXlsxExport
-    pim_connector.job.job_parameters.default_values_provider.product_model_csv_export.class:  Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModelCsvExport
-    pim_connector.job.job_parameters.default_values_provider.product_model_xlsx_export.class: Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModelCsvExport
 
     pim_connector.job.job_parameters.default_values_provider.simple_csv_import.class:         Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleCsvImport
     pim_connector.job.job_parameters.default_values_provider.simple_xlsx_import.class:        Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleXlsxImport
     pim_connector.job.job_parameters.default_values_provider.simple_yaml_import.class:        Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\SimpleYamlImport
-    pim_connector.job.job_parameters.default_values_provider.product_csv_import.class:        Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductCsvImport
-    pim_connector.job.job_parameters.default_values_provider.product_model_csv_import.class:  Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModelCsvImport
-    pim_connector.job.job_parameters.default_values_provider.product_xlsx_import.class:       Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductCsvImport
-    pim_connector.job.job_parameters.default_values_provider.product_model_xlsx_import.class: Pim\Component\Connector\Job\JobParameters\DefaultValuesProvider\ProductModelCsvImport
 
 services:
     # Default Parameters for Exports ----------------------------------------------------------------------------------
@@ -55,50 +47,6 @@ services:
         tags:
             - { name: akeneo_batch.job.job_parameters.default_values_provider }
 
-    pim_connector.job.job_parameters.default_values_provider.product_csv_export:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_csv_export.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_csv_export'
-            - '@pim_catalog.repository.channel'
-            - '@pim_catalog.repository.locale'
-            -
-                - 'csv_product_export'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_xlsx_export:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_xlsx_export.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_xlsx_export'
-            - '@pim_catalog.repository.channel'
-            - '@pim_catalog.repository.locale'
-            -
-                - 'xlsx_product_export'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_model_csv_export:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_model_csv_export.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_csv_export'
-            - '@pim_catalog.repository.channel'
-            - '@pim_catalog.repository.locale'
-            -
-                - 'csv_product_model_export'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_model_xlsx_export:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_model_xlsx_export.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_xlsx_export'
-            - '@pim_catalog.repository.channel'
-            - '@pim_catalog.repository.locale'
-            -
-                - 'xlsx_product_model_export'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
     # Default Parameters for Imports ----------------------------------------------------------------------------------
     pim_connector.job.job_parameters.default_values_provider.simple_csv_import:
         class: '%pim_connector.job.job_parameters.default_values_provider.simple_csv_import.class%'
@@ -135,41 +83,5 @@ services:
                 - 'xlsx_attribute_group_import'
                 - 'xlsx_currency_import'
                 - 'xlsx_group_type_import'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_csv_import:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_csv_import.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_csv_import'
-            -
-                - 'csv_product_import'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_xlsx_import:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_xlsx_import.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_xlsx_import'
-            -
-                - 'xlsx_product_import'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_model_csv_import:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_model_csv_import.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_csv_import'
-            -
-                - 'csv_product_model_import'
-        tags:
-            - { name: akeneo_batch.job.job_parameters.default_values_provider }
-
-    pim_connector.job.job_parameters.default_values_provider.product_model_xlsx_import:
-        class: '%pim_connector.job.job_parameters.default_values_provider.product_model_xlsx_import.class%'
-        arguments:
-            - '@pim_connector.job.job_parameters.default_values_provider.simple_xlsx_import'
-            -
-                - 'xlsx_product_model_import'
         tags:
             - { name: akeneo_batch.job.job_parameters.default_values_provider }
